requirements.txt
Django==2.1
django-js-asset==1.1.0
django-mptt==0.9.1
djangorestframework==3.8.2
psycopg2==2.7.5
pytz==2018.5
six==1.11.0
